+++ app/Http/Livewire/User/Components/RoleTable.php
@@ -10,6 +10,9 @@ use Spatie\Permission\Models\Role;
 class RoleTable extends DataTableComponent
 {
     protected $model = Role::class;
+    protected $listeners = [
+        'deleteConfirmed'
+    ];
 
     public function configure(): void
     {
@@ -32,23 +35,31 @@ class RoleTable extends DataTableComponent
                 ->searchable()
                 ->sortable(),
             Column::make(__('Created at'), 'created_at')
-                ->format(fn($value) => format_date($value))
+                ->format(fn ($value) => format_date($value))
                 ->sortable(),
             Column::make(__('Actions'), 'id')
                 ->view('livewire.user.components.role-action-menu'),
         ];
     }
 
-    public function actionEdit($id) {
+    public function actionEdit($id)
+    {
         return redirect()->to(route('role.edit', $id));
     }
 
-    public function actionDelete($id) {
-        $role = Role::where('id', $id)->first();
-        if ($role) {
-            $role->permissions()->detach();
-            $role->delete();
+    public function actionDelete($id)
+    {
+        $this->emitTo('components.delete-confirm-modal', 'deleteConfirmation', 'user.components.role-table', $id);
+    }
+
+    public function deleteConfirmed($id)
+    {
+        if ($id) {
+            $role = Role::where('id', $id)->first();
+            if ($role) {
+                $role->permissions()->detach();
+                $role->delete();
+            }
         }
-        return redirect()->to(route('role.index'));
     }
 }
